test/conftest: Introduce new database fixtures
They are essentially the same fixtures as before but with a better naming and documentation.
